@@ -16,10 +16,9 @@
 		"issues": "https://github.com/Trejjam/Utils/issues"
 	},
 	"require": {
-		"php": ">=7.1",
+		"php": "^8.1 | ^8.2",
 
-		"nette/di": "~3.0",
-		"trejjam/base-extension": "~0.7"
+		"nette/di": "~3.0"
 	},
 	"autoload": {
 		"psr-4": {
@@ -31,7 +30,7 @@
 	},
 	"extra": {
 		"branch-alias": {
-			"dev-master": "v2.1.x-dev"
+			"dev-master": "v3.0.x-dev"
 		}
 	}
 }
