@@ -6,86 +6,74 @@ use Tracy;
 
 class BaseTracyPanel implements Tracy\IBarPanel
 {
-	protected $count = 0;
+    protected int $count = 0;
 
-	/**
-	 * @var array
-	 */
-	protected $cmd = [];
+    protected array $cmd = [];
 
-	/**
-	 * @var string
-	 */
-	public $name;
+    public string $name;
 
-	public $disabled = FALSE;
+    public bool $disabled = FALSE;
 
-	/**
-	 * @var string
-	 */
-	protected $tabTemplate = '';
-	/**
-	 * @var string
-	 */
-	protected $panelTemplate = '';
+    protected string $tabTemplate = '';
+    protected string $panelTemplate = '';
 
-	public function __construct(AShellExecute $execute, $tabTemplate = NULL, $panelTemplate = NULL)
-	{
-		$execute->logger[] = [$this, 'logCmd'];
-		$this->name = $execute->getLoggerName();
+    public function __construct(AShellExecute $execute, string|null $tabTemplate = NULL, string|null $panelTemplate = NULL)
+    {
+        $execute->logger[] = [$this, 'logCmd'];
+        $this->name = $execute->getLoggerName();
 
-		if ( !is_null($tabTemplate)) {
-			$this->tabTemplate = $tabTemplate;
-		}
-		if ( !is_null($panelTemplate)) {
-			$this->panelTemplate = $panelTemplate;
-		}
-	}
+        if (!is_null($tabTemplate)) {
+            $this->tabTemplate = $tabTemplate;
+        }
+        if (!is_null($panelTemplate)) {
+            $this->panelTemplate = $panelTemplate;
+        }
+    }
 
-	public function logCmd(AShellExecute $that, $message, $title, $time)
-	{
-		if ($this->disabled) {
-			return;
-		}
-		$this->count++;
+    public function logCmd(AShellExecute $that, $message, $title, $time)
+    {
+        if ($this->disabled) {
+            return;
+        }
+        $this->count++;
 
-		$this->cmd[] = [
-			'message' => $message,
-			'title'   => $title,
-			'time'    => $time,
-		];
-	}
+        $this->cmd[] = [
+            'message' => $message,
+            'title' => $title,
+            'time' => $time,
+        ];
+    }
 
-	public function getTab()
-	{
-		$name = $this->name;
-		$count = $this->count;
+    public function getTab()
+    {
+        $name = $this->name;
+        $count = $this->count;
 
-		if ( !$count) {
-			return '';
-		}
+        if (!$count) {
+            return '';
+        }
 
-		ob_start();
-		require $this->tabTemplate;
+        ob_start();
+        require $this->tabTemplate;
 
-		return ob_get_clean();
-	}
+        return ob_get_clean();
+    }
 
 
-	public function getPanel()
-	{
-		$this->disabled = TRUE;
-		if ( !$this->count) {
-			return NULL;
-		}
+    public function getPanel()
+    {
+        $this->disabled = TRUE;
+        if (!$this->count) {
+            return NULL;
+        }
 
-		$name = $this->name;
-		$count = $this->count;
-		$cmd = $this->cmd;
+        $name = $this->name;
+        $count = $this->count;
+        $cmd = $this->cmd;
 
-		ob_start();
-		require $this->panelTemplate;
+        ob_start();
+        require $this->panelTemplate;
 
-		return ob_get_clean();
-	}
+        return ob_get_clean();
+    }
 }
