@@ -6,5 +6,5 @@ use Trejjam;
 
 interface IListingFactory
 {
-	function create(Trejjam\Utils\Helpers\IBaseList $list = NULL) : ListingFactory;
+    function create(Trejjam\Utils\Helpers\IBaseList|null $list = null): ListingFactory;
 }
