@@ -3,27 +3,28 @@ declare(strict_types=1);
 
 namespace Trejjam\Utils\DI;
 
-use Nette\DI\Definitions\FactoryDefinition;
+use Nette\DI\CompilerExtension;
 use Trejjam;
+use Nette\DI\Definitions\FactoryDefinition;
 
-class LatteExtension extends Trejjam\BaseExtension\DI\BaseExtension
+final class LatteExtension extends CompilerExtension
 {
-	protected $classesDefinition = [
-		'filter.json' => Trejjam\Utils\Latte\Filter\Json::class,
-		'filter.md5'  => Trejjam\Utils\Latte\Filter\Md5::class,
-		'filter.sha1' => Trejjam\Utils\Latte\Filter\Sha1::class,
-	];
+    public function beforeCompile()
+    {
+        $builder = $this->getContainerBuilder();
 
-	public function beforeCompile()
-	{
-		$builder = $this->getContainerBuilder();
+        $builder->addDefinition($this->prefix('filter.json'))
+            ->setType(Trejjam\Utils\Latte\Filter\Json::class);
+        $builder->addDefinition($this->prefix('filter.md5'))
+            ->setType(Trejjam\Utils\Latte\Filter\Md5::class);
+        $builder->addDefinition($this->prefix('filter.sha1'))
+            ->setType(Trejjam\Utils\Latte\Filter\Sha1::class);
 
-		/** @var FactoryDefinition $latteFactoryDefinition */
-		$latteFactoryDefinition = $builder->getDefinition('latte.latteFactory');
-		$latteFactory = $latteFactoryDefinition->getResultDefinition();
-
-		$latteFactory->addSetup('addFilter', ['json', [$this->prefix('@filter.json'), 'filter']]);
-		$latteFactory->addSetup('addFilter', ['md5', [$this->prefix('@filter.md5'), 'filter']]);
-		$latteFactory->addSetup('addFilter', ['sha1', [$this->prefix('@filter.sha1'), 'filter']]);
-	}
+        /** @var FactoryDefinition $latteFactoryDefinition */
+        $latteFactoryDefinition = $builder->getDefinition('latte.latteFactory');
+        $latteFactoryDefinition->getResultDefinition()
+            ->addSetup('addFilter', ['json', [$this->prefix('@filter.json'), 'filter']])
+            ->addSetup('addFilter', ['md5', [$this->prefix('@filter.md5'), 'filter']])
+            ->addSetup('addFilter', ['sha1', [$this->prefix('@filter.sha1'), 'filter']]);
+    }
 }
