@@ -7,12 +7,12 @@ use Nextras;
 
 class Rule
 {
-	const IPV4 = '(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)';
-	const IPV6 = '(?:[a-fA-F0-9]{1,4}:){7}[a-fA-F0-9]{1,4}';
-	const MAC  = '([a-fA-F0-9]{2}[:.-]?){7}[[a-fA-F0-9]{2}';
+	public const IPV4 = '(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)';
+	public const IPV6 = '(?:[a-fA-F0-9]{1,4}:){7}[a-fA-F0-9]{1,4}';
+	public const MAC  = '([a-fA-F0-9]{2}[:.-]?){7}[[a-fA-F0-9]{2}';
 
-	const DateTimeCompare = 'Trejjam\Utils\Helpers\Form\Rule::dateTimeCompare';
-	const DateCompare     = 'Trejjam\Utils\Helpers\Form\Rule::dateCompare';
+	public const DateTimeCompare = 'Trejjam\Utils\Helpers\Form\Rule::dateTimeCompare';
+    public const DateCompare     = 'Trejjam\Utils\Helpers\Form\Rule::dateCompare';
 
 	static function dateTimeCompare(Nextras\Forms\Controls\DateTimePicker $item, array $arg)
 	{
